requirements.txt
py-clob-client>=0.14.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0
